Don't pretty-print long queries by default
It seems that formatting very large queries with sqlparse and the indent
function takes a very long time with large query statements, so only try
to format query statements that are less than or equal to 2k characters
by default.  This can be overridden with an optional parameter to
show_queries.